Move 'posts' reducer to the separate file
Add postId as the DELETE_COMMENT param
