+++ src/redux/reducers.js
@@ -3,7 +3,7 @@ import _ from 'lodash';
 import { LOCATION_CHANGE } from 'react-router-redux';
 import { combineReducers } from 'redux';
 
-import { userParser, postParser, getSummaryPeriod } from '../utils';
+import { userParser, getSummaryPeriod } from '../utils';
 import { getToken } from '../services/auth';
 import { parseQuery } from '../utils/search-highlighter';
 import { formatDateFromShortString } from '../utils/get-date-from-short-string';
@@ -637,385 +637,7 @@ export function postsViewState(state = {}, action) {
   return state;
 }
 
-function updatePostData(state, action) {
-  const postId = action.payload.posts.id;
-  return { ...state, [postId]: postParser(action.payload.posts) };
-}
-
-const savePostStatusesReducer = asyncStatesMap(ActionTypes.SAVE_POST, {
-  getKey: getKeyBy('postId'),
-  keyMustExist: true,
-  applyState: (post, savePostStatus) => ({ ...post, savePostStatus }),
-});
-
-export function posts(state = {}, action) {
-  if (ActionHelpers.isFeedResponse(action)) {
-    return mergeByIds(state, (action.payload.posts || []).map(postParser));
-  }
-
-  // Handle the savePostStatus changes
-  state = savePostStatusesReducer(state, action);
-
-  switch (action.type) {
-    case response(ActionTypes.SHOW_MORE_COMMENTS): {
-      const post = state[action.payload.posts.id];
-      if (!post) {
-        return state;
-      }
-      const newPost = postParser(action.payload.posts);
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          omittedComments: newPost.omittedComments,
-          omittedCommentsOffset: newPost.omittedCommentsOffset,
-          comments: newPost.comments,
-        },
-      };
-    }
-    case response(ActionTypes.SHOW_MORE_LIKES_ASYNC): {
-      const post = state[action.payload.posts.id];
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          omittedLikes: 0,
-          likes: action.payload.posts.likes,
-        },
-      };
-    }
-    case response(ActionTypes.SAVE_EDITING_POST): {
-      const post = state[action.payload.posts.id];
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          body: action.payload.posts.body,
-          updatedAt: action.payload.posts.updatedAt,
-          attachments: action.payload.posts.attachments || [],
-          postedTo: action.payload.posts.postedTo,
-        },
-      };
-    }
-    case response(ActionTypes.DELETE_COMMENT): {
-      const { commentId } = action.request;
-      const post = Object.values(state).find((p) => p.comments?.includes(commentId));
-      if (!post) {
-        return state;
-      }
-
-      let { omittedCommentsOffset } = post;
-      const idx = post.comments.indexOf(commentId);
-      if (idx >= 0 && idx < post.omittedCommentsOffset) {
-        // If the deletion is before the omittedComments span
-        omittedCommentsOffset--;
-      }
-
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          comments: post.comments.filter((c) => c !== commentId),
-          omittedCommentsOffset,
-        },
-      };
-    }
-    case ActionTypes.REALTIME_COMMENT_DESTROY: {
-      const { postId, commentId } = action;
-      const post = state[postId];
-      if (!post || (!post.comments.includes(commentId) && post.omittedComments === 0)) {
-        return state;
-      }
-
-      return patchObjectByKey(state, postId, (post) => {
-        const p = { ...post };
-        const idx = p.comments.indexOf(commentId);
-        if (idx >= 0) {
-          if (idx < p.omittedCommentsOffset) {
-            p.omittedCommentsOffset--;
-          }
-          p.comments = p.comments.filter((c) => c !== commentId);
-        } else {
-          p.omittedComments = p.omittedComments - 1;
-          if (p.omittedComments <= 0) {
-            p.omittedComments = 0;
-            p.omittedCommentsOffset = 0;
-          }
-        }
-        return p;
-      });
-    }
-    case response(ActionTypes.ADD_COMMENT): {
-      const post = state[action.request.postId];
-      if (!post || post.comments?.includes(action.payload.comments.id)) {
-        return state;
-      }
-      return patchObjectByKey(state, action.request.postId, (post) => ({
-        ...post,
-        comments: [...(post.comments || []), action.payload.comments.id],
-      }));
-    }
-
-    // Likes
-
-    case ActionTypes.CLEAN_LIKE_ERROR: {
-      const { postId } = action;
-      const post = state[postId];
-      if (!post || !post.likeError) {
-        return state;
-      }
-      return {
-        ...state,
-        [postId]: {
-          ...post,
-          likeError: null,
-        },
-      };
-    }
-
-    case ActionTypes.LIKE_POST_OPTIMISTIC: {
-      const { postId, userId } = action.payload;
-      const post = state[postId];
-      if (_.includes(post.likes, userId)) {
-        return state;
-      }
-      return {
-        ...state,
-        [postId]: {
-          ...post,
-          likeError: null,
-          likes: [userId, ...post.likes],
-        },
-      };
-    }
-    case fail(ActionTypes.LIKE_POST): {
-      const { postId, userId } = action.request;
-      const post = state[postId];
-      if (!_.includes(post.likes, userId)) {
-        return state;
-      }
-      return {
-        ...state,
-        [postId]: {
-          ...post,
-          likeError: `Cannot like post: ${action.payload.err}`,
-          likes: _.without(post.likes, userId),
-        },
-      };
-    }
-
-    case ActionTypes.UNLIKE_POST_OPTIMISTIC: {
-      const { postId, userId } = action.payload;
-      const post = state[postId];
-      if (!_.includes(post.likes, userId)) {
-        return state;
-      }
-      return {
-        ...state,
-        [postId]: {
-          ...post,
-          likeError: null,
-          likes: _.without(post.likes, userId),
-        },
-      };
-    }
-    case fail(ActionTypes.UNLIKE_POST): {
-      const { postId, userId } = action.request;
-      const post = state[postId];
-      if (_.includes(post.likes, userId)) {
-        return state;
-      }
-      return {
-        ...state,
-        [postId]: {
-          ...post,
-          likeError: `Cannot unlike post: ${action.payload.err}`,
-          likes: [userId, ...post.likes],
-        },
-      };
-    }
-
-    case ActionTypes.REALTIME_LIKE_NEW: {
-      const {
-        postId,
-        users: [{ id: userId }],
-      } = action;
-      const post = state[postId];
-      if (!post && !action.post) {
-        return state;
-      }
-
-      const postToAct = post || postParser(action.post.posts);
-
-      if (postToAct.likes && _.includes(postToAct.likes, userId)) {
-        return {
-          ...state,
-          [postToAct.id]: postToAct,
-        };
-      }
-
-      const likes = action.iLiked
-        ? [postToAct.likes[0], userId, ...postToAct.likes.slice(1)]
-        : [userId, ...(postToAct.likes || [])];
-
-      return {
-        ...state,
-        [postToAct.id]: { ...postToAct, likes },
-      };
-    }
-
-    case ActionTypes.REALTIME_LIKE_REMOVE: {
-      const { postId, userId, isLikeVisible } = action;
-      const post = state[postId];
-      if (!post) {
-        return state;
-      }
-
-      const likes = _.without(post.likes, userId);
-      const omittedLikes = isLikeVisible ? post.omittedLikes : Math.max(0, post.omittedLikes - 1);
-
-      return {
-        ...state,
-        [post.id]: { ...post, likes, omittedLikes },
-      };
-    }
-
-    case response(ActionTypes.HIDE_POST): {
-      const post = state[action.request.postId];
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          isHidden: true,
-        },
-      };
-    }
-    case ActionTypes.REALTIME_POST_HIDE: {
-      const post = state[action.postId];
-      if (!post) {
-        return state;
-      }
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          isHidden: true,
-        },
-      };
-    }
-    case response(ActionTypes.UNHIDE_POST): {
-      const post = state[action.request.postId];
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          isHidden: false,
-        },
-      };
-    }
-    case ActionTypes.REALTIME_POST_UNHIDE: {
-      const post = state[action.postId];
-      if (!post) {
-        return state;
-      }
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          isHidden: false,
-        },
-      };
-    }
-
-    case response(ActionTypes.SAVE_POST): {
-      return patchObjectByKey(state, action.request.postId, (post) => ({
-        ...post,
-        isSaved: action.request.save,
-      }));
-    }
-
-    case ActionTypes.REALTIME_POST_SAVE: {
-      return patchObjectByKey(state, action.payload.postId, (post) => ({
-        ...post,
-        isSaved: action.payload.save,
-      }));
-    }
-
-    case response(ActionTypes.DISABLE_COMMENTS): {
-      const post = state[action.request.postId];
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          commentsDisabled: true,
-        },
-      };
-    }
-    case response(ActionTypes.ENABLE_COMMENTS): {
-      const post = state[action.request.postId];
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          commentsDisabled: false,
-        },
-      };
-    }
-    case response(ActionTypes.CREATE_POST): {
-      return updatePostData(state, action);
-    }
-    case response(ActionTypes.GET_SINGLE_POST): {
-      return updatePostData(state, action);
-    }
-    case ActionTypes.REALTIME_POST_NEW: {
-      return { ...state, [action.post.id]: postParser(action.post) };
-    }
-    case ActionTypes.REALTIME_POST_UPDATE: {
-      const post = state[action.post.id];
-      if (!post) {
-        return state;
-      }
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          body: action.post.body,
-          updatedAt: action.post.updatedAt,
-          attachments: action.post.attachments || [],
-          postedTo: action.post.postedTo,
-        },
-      };
-    }
-    case ActionTypes.REALTIME_COMMENT_NEW: {
-      const post = state[action.comment.postId];
-      if (!post) {
-        if (!action.post) {
-          return state;
-        }
-
-        return {
-          ...state,
-          [action.post.posts.id]: postParser(action.post.posts),
-        };
-      }
-      if (post.comments?.includes(action.comment.id)) {
-        return state;
-      }
-      return {
-        ...state,
-        [post.id]: {
-          ...post,
-          comments: [...(post.comments || []), action.comment.id],
-        },
-      };
-    }
-    case ActionTypes.UNAUTHENTICATED: {
-      return {};
-    }
-  }
-
-  return state;
-}
+export { posts } from './reducers/posts';
 
 export const postHideStatuses = asyncStatesMap(
   [
